Cambios en migracion

+++ BackendBolsaDeTrabajoUTN/BackendBolsaDeTrabajoUTN/Entities/Company.cs
@@ -13,14 +13,14 @@ namespace BackendBolsaDeTrabajoUTN.Entities
         public string CompanyAddress { get; set; }
         public string CompanyLocation { get; set; }
         public string CompanyPostalCode { get; set; }
-        public int CompanyPhone { get; set; }
+        public long CompanyPhone { get; set; }
         public string CompanyWebPage { get; set; }
 
         // datos de contacto
         public string CompanyPersonalName { get; set; }
         public string CompanyPersonalSurname { get; set; }
         public string CompanyPersonalJob { get; set; }
-        public int CompanyPersonalPhone { get; set; }
+        public long CompanyPersonalPhone { get; set; }
         public string CompanyRelationContact { get; set; }
         public bool CompanyPendingConfirmation { get; set; }
